chore: refactored the Survey Builder
